backend/app/services/bill_processor.py
import os
import magic
import pandas as pd
import pdfplumber
import xmltodict
from typing import Dict, Any, List, Tuple
from datetime import datetime
from werkzeug.datastructures import FileStorage
from ..models import Bill, BillAudit, LinkedAccountMeter

class BillProcessor:
    ALLOWED_EXTENSIONS = {
        'application/pdf': '.pdf',
        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet': '.xlsx',
        'application/vnd.ms-excel': '.xls',
        'text/xml': '.xml',
        'application/xml': '.xml',
        'text/plain': '.txt'
    }

    def __init__(self, file: FileStorage, linked_account_meter_id: int):
        self.file = file
        self.linked_account_meter_id = linked_account_meter_id
        self.mime_type = magic.from_buffer(file.read(2048), mime=True)
        file.seek(0)  # Reset file pointer after reading

    def process(self) -> Tuple[Bill, List[BillAudit]]:
        """Process the bill file and return the created bill and its audits"""
        if self.mime_type not in self.ALLOWED_EXTENSIONS:
            raise ValueError(f"Unsupported file type: {self.mime_type}")

        # Extract bill data based on file type
        bill_data = self._extract_bill_data()
        
        # Create bill record
        bill = Bill(
            linked_account_meter_id=self.linked_account_meter_id,
            bill_date=bill_data['bill_date'],
            due_date=bill_data['due_date'],
            amount=bill_data['amount'],
            usage_amount=bill_data.get('usage_amount'),
            source_type=self._get_source_type(),
            file_path=self._save_file()
        )

        # Perform audits
        audits = self._perform_audits(bill_data)

        return bill, audits

    def _extract_bill_data(self) -> Dict[str, Any]:
        """Extract bill data based on file type"""
        if self.mime_type == 'application/pdf':
            return self._extract_from_pdf()
        elif self.mime_type in ['application/vnd.openxmlformats-officedocument.spreadsheetml.sheet', 'application/vnd.ms-excel']:
            return self._extract_from_excel()
        elif self.mime_type in ['text/xml', 'application/xml']:
            return self._extract_from_xml()
        else:
            return self._extract_from_text()

    def _extract_from_pdf(self) -> Dict[str, Any]:
        """Extract bill data from PDF using pdfplumber"""
        with pdfplumber.open(self.file) as pdf:
            # Implementation would depend on specific PDF layout
            # This is a simplified example
            text = ""
            for page in pdf.pages:
                text += page.extract_text()
            
            # Parse text to extract bill data
            # This would need to be customized based on the actual PDF format
            return {
                'bill_date': datetime.now(),  # Placeholder
                'due_date': datetime.now(),   # Placeholder
                'amount': 0.0,                # Placeholder
                'usage_amount': 0.0           # Placeholder
            }

    def _extract_from_excel(self) -> Dict[str, Any]:
        """Extract bill data from Excel file"""
        df = pd.read_excel(self.file)
        # Implementation would depend on specific Excel layout
        return {
            'bill_date': datetime.now(),  # Placeholder
            'due_date': datetime.now(),   # Placeholder
            'amount': 0.0,                # Placeholder
            'usage_amount': 0.0           # Placeholder
        }

    def _extract_from_xml(self) -> Dict[str, Any]:
        """Extract bill data from XML file"""
        xml_dict = xmltodict.parse(self.file.read())
        # Implementation would depend on specific XML schema
        return {
            'bill_date': datetime.now(),  # Placeholder
            'due_date': datetime.now(),   # Placeholder
            'amount': 0.0,                # Placeholder
            'usage_amount': 0.0           # Placeholder
        }

    def _extract_from_text(self) -> Dict[str, Any]:
        """Extract bill data from text file"""
        text = self.file.read().decode('utf-8')
        # Implementation would depend on specific text format
        return {
            'bill_date': datetime.now(),  # Placeholder
            'due_date': datetime.now(),   # Placeholder
            'amount': 0.0,                # Placeholder
            'usage_amount': 0.0           # Placeholder
        }

    def _perform_audits(self, bill_data: Dict[str, Any]) -> List[BillAudit]:
        """Perform various audits on the bill data"""
        audits = []
        
        # Check if amount is positive
        amount_audit = BillAudit(
            audit_type='amount_validation',
            status='passed' if bill_data['amount'] > 0 else 'failed',
            message='Amount must be positive'
        )
        audits.append(amount_audit)

        # Check if dates are valid
        date_audit = BillAudit(
            audit_type='date_validation',
            status='passed' if bill_data['due_date'] > bill_data['bill_date'] else 'failed',
            message='Due date must be after bill date'
        )
        audits.append(date_audit)

        # Add more audits as needed

        return audits

    def _get_source_type(self) -> str:
        """Determine the source type based on file type"""
        if self.mime_type == 'application/pdf':
            return 'PDF'
        elif self.mime_type in ['application/vnd.openxmlformats-officedocument.spreadsheetml.sheet', 'application/vnd.ms-excel']:
            return 'Excel'
        elif self.mime_type in ['text/xml', 'application/xml']:
            return 'XML'
        else:
            return 'Text'

    def _save_file(self) -> str:
        """Save the file and return the file path"""
        # Implementation would depend on your file storage strategy
        # This is a placeholder
        return f"bills/{self.file.filename}"
