requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
boto3==1.28.36
pillow==10.0.0
pytesseract==0.3.10
pandas==2.1.0
numpy==1.25.2
gunicorn==21.2.0
xlrd==2.0.1
openpyxl==3.1.2
pdfplumber==0.10.2
python-dateutil==2.8.2
reportlab==4.0.4
xmltodict==0.13.0
requests==2.31.0
python-magic==0.4.27
SQLAlchemy-Utils==0.41.1
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
celery==5.3.4
redis==5.0.0
flower==2.0.1
Flask-Babel==4.0.0
pytz==2023.3
